Offloaded the call to gpu-info to a celery queue, to prevent a delay in the UI when switching tabs.
